(Viewer.py): Added stuff to read the history and priorities stuff from style sheets.  Added stuff so that links get rendered in 'atemp' style as the link is being followed.
(Browser.py): Removed age-old nit by binding Control-u to clear
everything to the left cursor in the URL entry field.

(AppletHTMLParser.py): modifications to the class to facilitate
changing the style of tags that have already been viewed (HISTORY).

(DefaultStylesheet.py): added styles for 'ahist' which are tags for
links that have already been viewed (HISTORY), and 'atemp' which are
tags for links that are in the processes of being followed.  Also
added priorities so that in general tag display priorities can be set
in the stylesheet.

(OpenURIDialog.py): string.strip the URI entered into the field