Relocated all the underlying framework stuff from PrefsDialogs.
Using a new initialization interface which eliminates reference to the
app and frame in the class interface.

Changed PrefsDialogMenu() to PrefsDialogsSetup(), with explicit
mention of the browser, etc.

Added a tk 'update' to the beginning of _poll_modified, for more
immediate user feedback.

Switching to identifying the second part of the preference name as the
"component" name.