New API for table cell geometry calculation of embedded windows.  If the object added to the viewer via add_subwindow() has an attribute called 'table_geometry', then that is a method that returns a 3 element tuple of the form:
(MINWIDTH, MAXWIDTH, HEIGHT)

where MINWIDTH is the narrowest the window can be, MAXWIDTH is the
widest the wndow can be, and HEIGHT is the tallest it can be.  These
are used when calculating layout for table cells.

If the object does *not* have this method, then an alternative method
is used whereby winfo_geometry() is called.  Note that Tk doesn't
always return the correct information, but that should be a rare
problem.  Current AppletFrame, Container (in table.py), ImageWindow
all have the table_geometry() method, but none of the widgets in
tktools.py does.

TBD: if an applet resizes its topmost frame, the table will not
resize.  It could if there was some mechanism to call back into
Table._notify().