Renamed DefaultStylesheet (module and class) to Stylesheet.
A single Stylesheet object is now on the app, and it registers a
pref callback to reload itself on changes to the relevant preference
groups.

Viewer.py continues to reconfigure itself on changes to those preference
groups, but leaves it to the stylesheet to reload itself (and relies on
the callback registration order to ensure that the stylesheet does so
before the reconfigure).