Implemented changes so that a mailto URL can be used as a form action.  This means mailto_access has to take (and know what to do with) an extra parameter.
Tried to emulate N$ behavior as close as makes sense.  This means
putting the form data in the body of the message, along with the
proper MIME headers (which are different if clicking on a mailto
URL -- this change handles both).

One gross hack.  Like N$, I wanted to put an X-URL: header in there
which is the URL of the page containing the mailto URL.  This turned
out to be difficult to get at at this point.  See upcoming Context.py
check-in for disgusting hack.