Proposed internal home page.  Suggestions are welcome.  Note that it has a gray background color, not white, to match the biglogo.gif image.
The license require this to acknowledge the copyright and trademark
status, as well as (pointers) to where the license can be obtained.
The inclusion of the title with the version number, the big logo image
(itself a link to the Grail home page) and the link to the Grail home
page, are my own inventions.  I've kept the use of <HR> or other
graphical whizz bang features to a minimum, to subtly suggest the
special status of this screen.  Perhaps it should contain more links
though (e.g. the same links as found in the Help menu, with some
enticing text...).

A following checkin will remove the splash screen and the dumb license
dialog.

Question: should the version displayed in the about.html file
accurately reflect the actual Grail __version__ number?  I don't have
an easy way to automate this inclusion...