Get rid of both the splash screen and the license dialog.
The splash is now part of the default home page, "grail:about.html".

The license is reachable via a link from there.

For users who change the default home page, both the "about" info and
the license are still available under the Help menu.  I hope this is
in compliance with the text of the license.

There's only one step left (which I'm not going to do right away): an
on-line registration page.  The about page should refer to this and
encourage users to register.  This requires a CGI script on our side
(separate from though similar to the one that hands out the sources).