I finally got fed up with the broken error handling and read the spec.  The following changes ensued.
1. Support version 1 of the protocol according to the on-line spec
(http://www.handle.net/docs/client_spec.html).  This only meant
reworking the error handling -- the error codes are all different now
(positive numbers).  Got rid of the code that attempted to strip
nonprinting characters from the error message -- instead, unpack the
fields of the error packet according to the table in the spec.

2. General cleanup -- revamped the main doc string to remove some
irrelevant musings and added references to the official protocol spec.
Revamped several other doc strings.  Removed the const_map -- it isn't
used.  Removed most references to the client library sources (which
are temporarily off line -- guess why :-).

I haven't been able to test this with Grail's hdlAPI yet (one annoying
bug in that code is that when an error is reported, and you click OK,
the Grail logo keeps spinning).

Other items to do are the config file and hash table parsing, which I
am unable to test.  See the TO DO list in the main doc string.