Fix a subtle bug that caused Python 1.5 to wrap lines in <PRE> tags and the like, even though the stylesheet specified wrap=none.  This didn't occur with Python 1.4.
A print statement in Viewer.configure_tags revealed that with Python
1.5, the cnf dictionary for the 'pre' tag was {'none': {}, 'wrap':
'none'}, while with Python 1.4 it was {'wrap': 'none', 'none': {}}.
Subtle, eh?  The different string hashing in 1.5 causes the order of
the keys in the dictionary to change.  I'm not sure why this confuses
the tag configure command, but apparently Tk only sees the first
option -- so with 1.5, the wrap option is not seen.

Anyway, the trace lead to the second call to Stylesheet .dictify_list:

    self.dictify_list(['styles', 'pre', 'wrap', 'none'])

Inspection of the logic of this function (no thanks to the doc string
:-( ) showed that this is really intended to be a fancy way to write

    self.attrs['styles']['pre']['wrap'] = 'none'

(with automatic insertion of missing dictionaries) and moreover that
there was logic bug that caused an additional entry 'none':{} to be
added to the final dictionary.

So, now that I've finally spent as much time writing this checkin
message as I've spent hunting down the bug, I'll commit.  Take it
away, Eric the orchestra leader!