(BookmarksController, BookmarksMenuViewer): A performance optimization.  Instead of rebuilding the bookmarks menu every time the menu is posted, only rebuild it if the internal model has changed. This means that the initial posting of the menu isn't any faster, but subsequent posts are very quick.
When the internal model changes `un-quietly' (i.e. other than an
expansion/collapse change), the menu has to be rebuilt.  I could try
to incrementally rebuild the menu, but the added complexity probably
isn't worth it.  Menu rebuilding `seems' to be about 1/3 of the time
it takes to actually parse the file which is only done on the initial
load.

Each browser window has it's own bookmarks menu, so each must rebuild
the menu individually, although they all use the same internal model.

Future Wish 1: It would be cool if the file were loaded while Grail
               was idle.

Future Wish 2: It would be cool if the menu was built while Grail was
               idle.

Future Wish 3: It would be cool if Grail occasionally stat'd the file
               looking for changes, a la NS.